Fix bugs, add new xkcd messages and improve xkcd display
Fix a bug with the daily messages system that would not schedule a new
message if the current day had no events.

Add new xkcd messages into the daily messages system, and improve the
xkcd messages by adding the alt text as a part of the message itself in
addition to being in the image.