Improve sleep command and add version-specific commands
Improve the sleep command to use the same mechanisms as the setEnabled
command. Add new wakeUp command that acts as an alias for setEnabled
true. Closes #34.

Add the ability to send a command to a specific version of LaTeX Bot
with the @LaTeX Bot <version> prefix, for example @LaTeX Bot v0.6.0 ping
would only cause version 0.6.0 to respond.