Improve and fix bugs in admin commands
Make admin commands output an error message when used in private DMs.

Fix a bug in moveMessages where it would crash when trying to handle
reactions and improve it to say where the moved messages came from, send
a message after it finishes and handle attachments correctly. (Update to
pyryver 0.3.3 for this.)