Temporary fix for #49
Temporary fix for issue #49, moveMessages attachment bug. Move the
message deletion to the end so the messages don't get eaten if the
moving fails. Add a try/catch around the attachment sending and fall
back to sending the message without attachments if it fails and report
the error. Add checks to make sure the combined message length stays
under 3900 characters.

This should temporarily resolve #11. Maybe moving the message deletion
to the end would fix it. In the worst case, moveMessages will only lose
attachments when it fails now.